Refactor move and delete operators into separate modules
Moved the move and delete asset operators from manage.py into dedicated move.py and delete.py modules for better organization and maintainability. Updated imports and registration in __init__.py and manage.py. Enhanced companion file and folder cleanup logic for both move and delete operations. Updated file_io.py and metadata.py to use shared datablock collection constants.
